Update on "Codegen EffectTokensWrapper in aot_autograd"
Replace the EffectTokensWrapper.post_compile closure with a codegen'd
function that bakes num_tokens as a compile-time constant. When
num_tokens == 0 (the common case), the wrapper early-returns the
compiled_fn directly, eliminating the wrapping closure entirely. When
num_tokens > 0, the generated function inlines the exact None-prepend
and output-slice with constant indices.

Generated code for 1 effect token:

    def _effect_tokens_wrapper(_compiled_fn_, args):
        args = [None, *args]
        outs = _compiled_fn_(args)
        if outs is None:
            return None
        return outs[1:]

EffectTokensWrapper step in isolation (us/call):

| Case | Before (closure) | After (codegen) | Speedup |
|---|---|---|---|
| 1 token, 5 args | 0.23 us | 0.19 us | 1.2x |
| 2 tokens, 5 args | 0.23 us | 0.19 us | 1.2x |
| 0 tokens, 5 args | 0.10 us | 0.04 us | 2.5x |

[ghstack-poisoned]
